pyproject.toml
[tool.poetry]
name = "property-finder"
version = "0.1.0"
description = ""
authors = ["aliseylaneh <aliseylaneh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "5.1.4"
django-environ = "0.11.2"
djangorestframework = "3.15.2"
django-celery-results = "2.5.1"
django-celery-beat = "2.7.0"
django-filter = "24.3"
django-extensions = "3.2.3"
django-cors-headers = "4.6.0"
drf-spectacular = "0.28.0"
django-redis = "5.4.0"
django-microservice-common = "0.2.1"
pydantic = "^2.10.4"
psycopg2-binary = "^2.9.10"
uvicorn = "^0.34.0"
elasticsearch-dsl = {extras = ["async"], version = "^8.17.1"}
kafka-python = "^2.0.2"
celery = {extras = ["redis"], version = "^5.4.0"}
django-elasticsearch-dsl = "^8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
